+++ app/src/main/java/com/example/playlistmaker/data/SearchHistory.kt
@@ -1,7 +1,8 @@
-package com.example.playlistmaker
+package com.example.playlistmaker.data
 
 import android.content.SharedPreferences
 import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.stringToTrackList
 import com.google.gson.Gson
 
 class SearchHistory(private val sharedPreferences: SharedPreferences) {
