+++ app/src/main/java/com/example/playlistmaker/presentation/ui/main/MainActivity.kt
@@ -7,10 +7,10 @@ import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
-import com.example.playlistmaker.data.SearchActivity
 import com.example.playlistmaker.databinding.ActivityMainBinding
 import com.example.playlistmaker.presentation.ui.library.LibraryActivity
 import com.example.playlistmaker.presentation.ui.settings.SettingsActivity
+import com.example.playlistmaker.presentation.ui.tracks.SearchActivity
 
 class MainActivity : AppCompatActivity() {
