+++ app/src/main/res/layout/activity_search.xml
@@ -6,7 +6,7 @@
     android:id="@+id/search"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".data.SearchActivity">
+    tools:context=".presentation.ui.tracks.SearchActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/toolbar"
